# openapi/api-gateway.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/gateway/auth/signup": {
      "post": {
        "description": "\nRegisters a new user under the provided **tenant** environment.\n\n### 🧩 Flow\n1. Validates **x-tenant-id** (must be provided in header).  \n2. Creates a new user with securely hashed password.  \n3. Automatically sends a **welcome email** (non-blocking).  \n4. Handles duplicate or validation errors gracefully.\n\n### ⚠️ Important Notes\n- The **x-tenant-id** header is mandatory to associate the user with the correct tenant.\n- Duplicate username/email/mobile entries will be rejected.\n- Passwords are securely hashed before storage.\n",
        "operationId": "AuthGatewayController_signup",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "User signup details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              },
              "examples": {
                "validExample": {
                  "summary": "Example signup request",
                  "value": {
                    "name": "John Doe",
                    "username": "johndoe",
                    "email": "john@example.com",
                    "password": "MySecurePassword@123",
                    "mobile": "01712345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account successfully created under tenant",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Your account has been created successfully. You can now log in to your AeroStitch workspace.",
                    "data": {
                      "status": "success",
                      "code": "USER_CREATED",
                      "data": {
                        "id": "uuid",
                        "name": "John Doe",
                        "username": "johndoe",
                        "email": "john@example.com",
                        "role": "user"
                      }
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (e.g., invalid email or weak password)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Signup failed due to invalid or missing information. Please review your inputs.",
                    "data": {
                      "status": "error",
                      "code": "VALIDATION_FAILED",
                      "message": "Password must be at least 6 characters"
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Duplicate username/email/mobile detected",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Signup failed: Some information already exists in the system. Please use different values.",
                    "data": {
                      "status": "error",
                      "code": "DUPLICATE_ENTRY",
                      "message": "Duplicate fields: email \"john@example.com\""
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error during signup",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Signup failed due to a system error. Please try again later.",
                    "data": {
                      "status": "error",
                      "code": "INTERNAL_ERROR",
                      "error": "Database connection lost"
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          }
        },
        "summary": "User Signup (Tenant-Aware Registration)",
        "tags": [
          "AuthGateway"
        ]
      }
    },
    "/gateway/auth/login": {
      "post": {
        "description": "\nInitiates the **OTP-based login flow** by validating credentials and sending a one-time password to the user’s registered email.\n\n### 🧩 Flow\n1. Validate tenant connection using **x-tenant-id** header.  \n2. Find user by **username/email/mobile**.  \n3. Verify password and account status.  \n4. Generate OTP (5-minute validity).  \n5. Send OTP via email.\n\n### ⚠️ Important Notes\n- This endpoint **does not log the user in immediately** — it sends an OTP for the next verification step.\n- Locked or suspended users cannot proceed.\n",
        "operationId": "AuthGatewayController_login",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "User credentials for login",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "examples": {
                "validExample": {
                  "summary": "Example login request",
                  "value": {
                    "usernameOrEmailOrMobile": "john.doe@example.com",
                    "password": "MySecurePassword@123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP successfully sent to registered email",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "A verification OTP has been sent to your registered email address. Please check your inbox.",
                    "data": {
                      "status": "success",
                      "code": "OTP_SENT",
                      "data": {
                        "loginId": "f5b2e8c3-8f4a-4a9c-9eab-8b23d67c2337",
                        "channel": "email",
                        "maskedEmail": "j***@example.com"
                      }
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid tenant or missing header",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Login failed: Tenant environment is not initialized. Please retry after selecting the correct workspace.",
                    "data": {
                      "status": "error",
                      "code": "TENANT_CONNECTION_MISSING",
                      "details": {
                        "tenantId": "tenant_12345"
                      }
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials, locked account, or wrong password",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "message": "Invalid credentials. Please check your username, email, or mobile number and try again.",
                        "data": {
                          "status": "error",
                          "code": "INVALID_CREDENTIALS",
                          "field": "usernameOrEmailOrMobile"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    },
                    {
                      "example": {
                        "message": "Your account is currently locked. Please check your email for unlock instructions or contact support.",
                        "data": {
                          "status": "error",
                          "code": "ACCOUNT_LOCKED",
                          "email": "john.doe@example.com"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    },
                    {
                      "example": {
                        "message": "Incorrect password. Please try again or reset your password if forgotten.",
                        "data": {
                          "status": "error",
                          "code": "INVALID_PASSWORD",
                          "field": "password"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "System or email failure",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "message": "We could not send the OTP to your email at this moment. Please try again later.",
                        "data": {
                          "status": "error",
                          "code": "OTP_SEND_FAILED",
                          "error": "SMTP connection timeout"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    },
                    {
                      "example": {
                        "message": "Login failed due to a system error. Please try again later.",
                        "data": {
                          "status": "error",
                          "code": "INTERNAL_ERROR",
                          "error": "Database connection lost"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "User Login (Initiate OTP-based authentication)",
        "tags": [
          "AuthGateway"
        ]
      }
    },
    "/gateway/auth/login/verify": {
      "post": {
        "description": "\nVerifies the OTP and issues **access and refresh tokens** for session-based authentication.\n\n### 🧩 Flow\n1. Validate tenant via **x-tenant-id** header.  \n2. Verify the OTP against the stored reference.  \n3. Issue **JWT tokens** (access & refresh).  \n4. Record session information.  \n5. Return tokens and user details.\n",
        "operationId": "AuthGatewayController_verifyOtp",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "OTP verification payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              },
              "examples": {
                "validExample": {
                  "summary": "Example request",
                  "value": {
                    "loginId": "8bfbec7f-3a2f-4d1f-a8e2-92ef1d2f3b77",
                    "otp": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified and tokens issued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "You have successfully logged in to AeroStitch.",
                    "data": {
                      "status": "success",
                      "code": "LOGIN_SUCCESS",
                      "data": {
                        "accessToken": "jwt-access-token",
                        "refreshToken": "jwt-refresh-token",
                        "sessionId": "uuid",
                        "user": {
                          "id": "uuid",
                          "username": "johndoe",
                          "role": "user"
                        }
                      }
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tenant missing or invalid environment",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "OTP verification failed: Tenant environment not initialized. Please retry.",
                    "data": {
                      "status": "error",
                      "code": "TENANT_CONNECTION_MISSING",
                      "details": {
                        "tenantId": "tenant_abc"
                      }
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "OTP invalid, expired, or mismatched",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "message": "Your OTP has expired or is invalid. Please request a new OTP to continue.",
                        "data": {
                          "status": "error",
                          "code": "OTP_EXPIRED_OR_INVALID",
                          "loginId": "8bfbec7f-3a2f-4d1f-a8e2-92ef1d2f3b77"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    },
                    {
                      "example": {
                        "message": "Incorrect OTP entered. Please check and try again.",
                        "data": {
                          "status": "error",
                          "code": "INVALID_OTP",
                          "field": "otp"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected system error during OTP verification",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Login verification failed due to an unexpected system error. Please try again later.",
                    "data": {
                      "status": "error",
                      "code": "INTERNAL_ERROR",
                      "error": "Redis connection lost"
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          }
        },
        "summary": "Verify OTP and issue JWT tokens",
        "tags": [
          "AuthGateway"
        ]
      }
    },
    "/gateway/auth/unlock": {
      "get": {
        "description": "\nUnlocks a user account using a token received via email after too many failed login attempts.\n\n### 🧩 Flow\n1. User receives an **unlock link** through email after repeated login failures.  \n2. The link contains a **token** (e.g., `https://yourapp.com/api/auth/unlock?token=abc123`).  \n3. The token is verified in Redis; if valid, the user account is unlocked and can log in again.  \n4. The unlock token is **deleted** immediately after successful verification.\n\n### ⚠️ Notes\n- The **x-tenant-id** header must always be provided.\n- The token expires automatically after a set duration (e.g., 30 minutes).\n- If the token is invalid, expired, or reused, an appropriate error response is returned.\n",
        "operationId": "AuthGatewayController_unlock",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "Unlock token from the email link",
            "schema": {
              "example": "df81b3e2-58d4-4a55-9b70-1fbd45a9f02e",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account unlocked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Your account has been unlocked successfully. You may now log in again.",
                    "data": {
                      "userId": "user_uuid",
                      "tenantId": "tenant_abc"
                    },
                    "meta": {
                      "status": "success",
                      "code": "ACCOUNT_UNLOCKED"
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid, expired, or incomplete unlock token",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "message": "The unlock link is invalid or has expired. Please request a new unlock email.",
                        "data": null,
                        "meta": {
                          "status": "error",
                          "code": "INVALID_OR_EXPIRED_TOKEN"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    },
                    {
                      "example": {
                        "message": "Unlock request failed due to incomplete token data. Please generate a new unlock link.",
                        "data": null,
                        "meta": {
                          "status": "error",
                          "code": "TOKEN_DATA_INCOMPLETE"
                        },
                        "ts": "2025-10-11T12:30:45.123Z"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error during unlock process",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Account unlock failed due to a system error. Please try again later.",
                    "data": null,
                    "meta": {
                      "status": "error",
                      "code": "INTERNAL_ERROR",
                      "error": "Redis connection lost"
                    },
                    "ts": "2025-10-11T12:30:45.123Z"
                  }
                }
              }
            }
          }
        },
        "summary": "Unlock Account (via Email Token)",
        "tags": [
          "AuthGateway"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "Public HTTP entrypoint routing to microservices (TCP).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://aero.darmist.com/backend",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:3501",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Use a valid **access token**.\n\nFormat: `Bearer <ACCESS_TOKEN>`",
        "in": "header"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}